#138 refactor: place로 통합함에 따른 변경
