+++ backend/src/main/java/meetween/backend/place/presentation/PlaceController.java
@@ -1,8 +1,8 @@
 package meetween.backend.place.presentation;
 
 import meetween.backend.place.application.PlaceService;
-import meetween.backend.place.dto.RestaurantResponse;
-import meetween.backend.place.dto.RestaurantsByLocationRequest;
+import meetween.backend.place.dto.response.RestaurantResponse;
+import meetween.backend.place.dto.request.PlacesByLocationRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,7 +17,7 @@ public class PlaceController {
     }
 
     @GetMapping("/near")
-    public ResponseEntity<RestaurantResponse> getNearRestaurants(@RequestBody final RestaurantsByLocationRequest request) {
+    public ResponseEntity<RestaurantResponse> getNearRestaurants(@RequestBody final PlacesByLocationRequest request) {
         RestaurantResponse response = placeService.getNearRestaurants(request);
         return ResponseEntity.ok().body(response);
     }
